requirements.txt
python-telegram-bot==21.5
Flask==3.1.2
google-generativeai==0.8.5
